// client/cms/src/type/I18nKeyType.ts
export interface I18nKey {
  signin: string;
  username: string;
  enter_email: string;
  email_required: string;
  email_invalid: string;
  password: string;
  enter_password: string;
  password_required: string;
  password_invalid: string;
  password_tooltip: string;
  remember: string;
  login: string;
  user: string;
  user_management: string;
  ID: string;
  name: string;
  grade: string;
  mobile_number: string;
  registration_date: string;
  subscription_start_date: string;
  total_subscription_period: string;
  coupon: string;
  remarks: string;
  no: string;
  brand_name: string;
  dashboard: string;
  journal_management: string;
  author_management: string;
  column_management: string;
  article_management: string;
  repository_management: string;
  handbooks_management: string;
  profile_management: string;
  glossary_management: string;
  shop_management: string;
  ads_management: string;
  quiz_poll_management: string;
  setting_management: string;
  logout: string;
  confirm_logout: string;
  page_not_found: string;
  back_to_homepage: string;
  email: string;
}

export enum TranslationKey {
  Signin = 'signin',
  Username = 'username',
  EnterEmail = 'enter_email',
  EmailRequired = 'email_required',
  EmailInvalid = 'email_invalid',
  Password = 'password',
  EnterPassword = 'enter_password',
  PasswordRequired = 'password_required',
  PasswordInvalid = 'password_invalid',
  PasswordTooltip = 'password_tooltip',
  Remember = 'remember',
  Login = 'login',
  User = 'user',
  UserManagement = 'user_management',
  ID = 'ID',
  Name = 'name',
  Grade = 'grade',
  MobileNumber = 'mobile_number',
  RegistrationDate = 'registration_date',
  SubscriptionStartDate = 'subscription_start_date',
  TotalSubscriptionPeriod = 'total_subscription_period',
  Coupon = 'coupon',
  Remarks = 'remarks',
  No = 'no',
  BrandName = 'brand_name',
  Dashboard = 'dashboard',
  JournalManagement = 'journal_management',
  AuthorManagement = 'author_management',
  ColumnManagement = 'column_management',
  ArticleManagement = 'article_management',
  RepositoryManagement = 'repository_management',
  HandbooksManagement = 'handbooks_management',
  ProfileManagement = 'profile_management',
  GlossaryManagement = 'glossary_management',
  ShopManagement = 'shop_management',
  ADSManagement = 'ads_management',
  QuizPollManagement = 'quiz_poll_management',
  SettingManagement = 'setting_management',
  Logout = 'logout',
  ConfirmLogout = 'confirm_logout',
  PageNotFound = 'page_not_found',
  BackToHomepage = 'back_to_homepage',
  Email = 'email',
}